Make Attribute Topology

given a set of lines with starting and ending nodes that form a geometric network, construct an attribute topology.

Usage

make_attribute_topology(x, min_distance)

# S3 method for data.frame
make_attribute_topology(x, min_distance)

# S3 method for hy
make_attribute_topology(x, min_distance)

Arguments

x

data.frame network compatible with hydroloom_names.

min_distance

numeric distance in units compatible with the units of the projection of lines. If no nodes are found within this distance, no connection will be returned.

Value

data.frame with id and toid

Details

If a future plan is set up, node distance calculations will be applied using future workers.

Examples


x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

y <- dplyr::select(x, COMID)

y <- sf::st_transform(y, 5070)

z <- make_attribute_topology(y, 10)
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~03s            |++                                                | 2 % ~06s            |++                                                | 3 % ~04s            |+++                                               | 4 % ~04s            |+++                                               | 5 % ~04s            |++++                                              | 6 % ~03s            |++++                                              | 7 % ~03s            |+++++                                             | 9 % ~03s            |+++++                                             | 10% ~03s            |++++++                                            | 11% ~03s            |++++++                                            | 12% ~03s            |+++++++                                           | 13% ~03s            |+++++++                                           | 14% ~03s            |++++++++                                          | 15% ~03s            |++++++++                                          | 16% ~03s            |+++++++++                                         | 17% ~02s            |++++++++++                                        | 18% ~02s            |++++++++++                                        | 19% ~02s            |+++++++++++                                       | 20% ~02s            |+++++++++++                                       | 21% ~02s            |++++++++++++                                      | 22% ~02s            |++++++++++++                                      | 23% ~02s            |+++++++++++++                                     | 24% ~02s            |+++++++++++++                                     | 26% ~02s            |++++++++++++++                                    | 27% ~02s            |++++++++++++++                                    | 28% ~02s            |+++++++++++++++                                   | 29% ~02s            |+++++++++++++++                                   | 30% ~02s            |++++++++++++++++                                  | 31% ~02s            |++++++++++++++++                                  | 32% ~02s            |+++++++++++++++++                                 | 33% ~02s            |++++++++++++++++++                                | 34% ~02s            |++++++++++++++++++                                | 35% ~02s            |+++++++++++++++++++                               | 36% ~02s            |+++++++++++++++++++                               | 37% ~02s            |++++++++++++++++++++                              | 38% ~02s            |++++++++++++++++++++                              | 39% ~02s            |+++++++++++++++++++++                             | 40% ~02s            |+++++++++++++++++++++                             | 41% ~02s            |++++++++++++++++++++++                            | 43% ~02s            |++++++++++++++++++++++                            | 44% ~02s            |+++++++++++++++++++++++                           | 45% ~02s            |+++++++++++++++++++++++                           | 46% ~01s            |++++++++++++++++++++++++                          | 47% ~01s            |++++++++++++++++++++++++                          | 48% ~01s            |+++++++++++++++++++++++++                         | 49% ~01s            |+++++++++++++++++++++++++                         | 50% ~01s            |++++++++++++++++++++++++++                        | 51% ~01s            |+++++++++++++++++++++++++++                       | 52% ~01s            |+++++++++++++++++++++++++++                       | 53% ~01s            |++++++++++++++++++++++++++++                      | 54% ~01s            |++++++++++++++++++++++++++++                      | 55% ~01s            |+++++++++++++++++++++++++++++                     | 56% ~01s            |+++++++++++++++++++++++++++++                     | 57% ~01s            |++++++++++++++++++++++++++++++                    | 59% ~01s            |++++++++++++++++++++++++++++++                    | 60% ~01s            |+++++++++++++++++++++++++++++++                   | 61% ~01s            |+++++++++++++++++++++++++++++++                   | 62% ~01s            |++++++++++++++++++++++++++++++++                  | 63% ~01s            |++++++++++++++++++++++++++++++++                  | 64% ~01s            |+++++++++++++++++++++++++++++++++                 | 65% ~01s            |+++++++++++++++++++++++++++++++++                 | 66% ~01s            |++++++++++++++++++++++++++++++++++                | 67% ~01s            |+++++++++++++++++++++++++++++++++++               | 68% ~01s            |+++++++++++++++++++++++++++++++++++               | 69% ~01s            |++++++++++++++++++++++++++++++++++++              | 70% ~01s            |++++++++++++++++++++++++++++++++++++              | 71% ~01s            |+++++++++++++++++++++++++++++++++++++             | 72% ~01s            |+++++++++++++++++++++++++++++++++++++             | 73% ~01s            |++++++++++++++++++++++++++++++++++++++            | 74% ~01s            |++++++++++++++++++++++++++++++++++++++            | 76% ~01s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~01s            |+++++++++++++++++++++++++++++++++++++++           | 78% ~01s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~01s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~01s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=03s  

x <- add_toids(hy(x), return_dendritic = FALSE)

x[x$id == x$id[1],]$toid
#> [1] 8894334
z[z$COMID == x$id[1],]$toid
#> [1] 8894334